[refactor] rename fragment
fragment들을 각 기능에 맞도록 네이밍 다시 진행
