+++ app/src/main/java/org/android/go/sopt/account/AccountFragment.kt
@@ -1,15 +1,14 @@
-package org.android.go.sopt.main
+package org.android.go.sopt.account
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import org.android.go.sopt.databinding.FragmentSearchBinding
-
-class SearchFragment : Fragment() {
-    private var _binding: FragmentSearchBinding? = null
-    private val binding: FragmentSearchBinding
+import org.android.go.sopt.databinding.FragmentAccountBinding
+class AccountFragment : Fragment() {
+    private var _binding: FragmentAccountBinding? = null
+    private val binding: FragmentAccountBinding
         get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
 
     override fun onCreateView(
@@ -17,7 +16,7 @@ class SearchFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSearchBinding.inflate(inflater, container, false)
+        _binding = FragmentAccountBinding.inflate(inflater, container, false)
         return binding.root
     }
