+++ app/src/main/java/org/android/go/sopt/playlist/ListFragment.kt
@@ -1,17 +1,16 @@
 package org.android.go.sopt.playlist
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.ConcatAdapter
-import org.android.go.sopt.databinding.FragmentHomeBinding
+import org.android.go.sopt.databinding.FragmentListBinding
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding: FragmentHomeBinding
+class ListFragment : Fragment() {
+    private var _binding: FragmentListBinding? = null
+    private val binding: FragmentListBinding
         get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
 
     override fun onCreateView(
@@ -19,7 +18,7 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentListBinding.inflate(inflater, container, false)
         return binding.root
     }
