+++ app/src/main/java/org/android/go/sopt/playlist/PlaylistActivity.kt
@@ -4,9 +4,9 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import org.android.go.sopt.R
-import org.android.go.sopt.account.GalleryFragment
+import org.android.go.sopt.account.AccountFragment
 import org.android.go.sopt.databinding.ActivityPlaylistBinding
-import org.android.go.sopt.main.SearchFragment
+import org.android.go.sopt.main.MainFragment
 
 class PlaylistActivity : AppCompatActivity() {
     private lateinit var binding: ActivityPlaylistBinding
@@ -17,13 +17,13 @@ class PlaylistActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         supportFragmentManager.findFragmentById(R.id.fcv_main)
-            ?: supportFragmentManager.beginTransaction().add(R.id.fcv_main, HomeFragment()).commit()
+            ?: supportFragmentManager.beginTransaction().add(R.id.fcv_main, ListFragment()).commit()
 
         binding.bnvMain.setOnItemSelectedListener { item ->
             when (item.itemId) {
-                R.id.menu_playlist -> changeFragment(HomeFragment())
-                R.id.menu_main -> changeFragment(SearchFragment())
-                R.id.menu_account -> changeFragment(GalleryFragment())
+                R.id.menu_playlist -> changeFragment(ListFragment())
+                R.id.menu_main -> changeFragment(MainFragment())
+                R.id.menu_account -> changeFragment(AccountFragment())
             }
             true
         }
@@ -32,7 +32,7 @@ class PlaylistActivity : AppCompatActivity() {
             if (item.itemId == R.id.menu_playlist) {
                 when (val currentFragment =
                     supportFragmentManager.findFragmentById(R.id.fcv_main)) {
-                    is HomeFragment -> {
+                    is ListFragment -> {
                         currentFragment.scrollToTop()
                     }
                 }
